refactor(#38): 공통 wrapper 컴포넌트 위치 변경
