+++ libs/ui/src/lib/pages/index.ts
@@ -6,7 +6,8 @@ export * from './SummaryPage';
 export * from './ChatRoomPage';
 export * from './ServerProfilePage';
 export * from './LoginPage';
-export * from './wrapper/pages/MainRoomWrapper';
+export * from './layout/MainRoomWrapper';
 export * from './SignUpPage';
-export * from './wrapper/pages/ServerProfileWrapper';
-export * from './wrapper';
+export * from './layout/ServerProfileWrapper';
+// export * from './wrapper';
+export * from './layout';
