+++ libs/ui/src/lib/pages/layout/index.ts
@@ -1,5 +1,6 @@
-'use client';
-
 export * from './MainLayoutWrapper';
 export * from './RootLayoutWrapper';
 export * from './ServerIdLayoutWrapper';
+export * from './MainRoomWrapper';
+export * from './ServerProfileWrapper';
+export * from './MainPageWrapper';
