--- libs/ui/src/lib/pages/wrapper/pages/index.ts
@@ -1,3 +1,0 @@
-export * from './MainPageWrapper';
-export * from './MainRoomWrapper';
-export * from './ServerProfileWrapper';
